+++ p5-analysis/src/commands/library-commands.ts
@@ -5,13 +5,17 @@ import { exec } from 'child_process';
 
 nunjucks.configure(`${__dirname}/templates`, { autoescape: false });
 
-export function describeLibrary(name: string) {
+export function describeLibrary(name: string, { json = false }) {
   const library = Library.find({ name });
   if (!library) {
     console.warn(`Library ${name} not found`);
     process.exit(1);
   }
-  console.log(nunjucks.render('library.njk', { library }));
+  if (json) {
+    console.log(JSON.stringify(library, null, 2));
+  } else {
+    console.log(nunjucks.render('library.njk', { library }));
+  }
 }
 
 export function printLibraryProperty(
@@ -32,17 +36,21 @@ export function printLibraryProperty(
   );
 }
 
-export function listLibraries({ verbose = false }) {
-  if (!verbose) {
-    console.log(Library.all.map(l => l.name).join('\n'));
-    return;
+export function listLibraries({ json = false, verbose = false }) {
+  {
+    if (json) {
+      console.log(JSON.stringify(Library.all, null, 2));
+    } else if (verbose) {
+      console.log(
+        nunjucks.render('list-libraries.njk', {
+          libraries: Library.all,
+          categories: Library.categories
+        })
+      );
+    } else {
+      console.log(Library.all.map(l => l.name).join('\n'));
+    }
   }
-  console.log(
-    nunjucks.render('list-libraries.njk', {
-      libraries: Library.all,
-      categories: Library.categories,
-    })
-  );
 }
 
 export async function updateDescriptions() {
@@ -54,7 +62,7 @@ export async function updateDescriptions() {
           exec(
             `npm view --json ${lib.packageName} description`,
             {
-              encoding: 'utf-8',
+              encoding: 'utf-8'
             },
             (error, stdout) => (error ? reject(error) : resolve(JSON.parse(stdout)))
           )